Code Cleanup | Home Page Navigation Label
