+++ smart_dog_app/package.json
@@ -11,6 +11,7 @@
     "csstools": "^0.1.5",
     "dotenv": "^8.1.0",
     "eslint": "^6.3.0",
+    "eslint-plugin-react-hooks": "^2.2.0",
     "json-server": "^0.15.1",
     "package.json": "^2.0.1",
     "react": "^16.9.0",
