+++ smart_dog_app/src/App.js
@@ -2,10 +2,8 @@ import React from "react";
 import {useAuth0} from "./auth0-wrapper";
 import {Layout, Header, Navigation, Drawer, Content } from 'react-mdl';
 import { Link } from 'react-router-dom';
-import Main from './components/main';
-import Login from "./components/login";
-
-require('dotenv').config();
+import Main from './main';
+import Login from "./components/Login_Out";
 
 function App() {
     const {loading} = useAuth0;
@@ -19,7 +17,7 @@ function App() {
     return (
         <div className="demo-big-content">
             <Layout fixedHeader>
-                <Header className="header-color" title="SMARTDog" scroll>
+                <Header className="header-color" title={<Link to={'/'} style={{color: 'white'}}>SMARTDog</Link>} scroll>
                     <Navigation>
                         <Link to="/about">About</Link>
                         <Link to="/works">How It Works</Link>
@@ -27,7 +25,7 @@ function App() {
                         <Login/>
                     </Navigation>
                 </Header>
-                <Drawer title="SMARTDog">
+                <Drawer title = {<Link to={'/'} style={{color: 'lightslategray'}}>SMARTDog</Link>}>
                     <Navigation>
                         <Link to="/about">About</Link>
                         <Link to="/works">How It Works</Link>
