+++ smart_dog_app/src/components/Login_Out.jsx
@@ -1,0 +1,31 @@
+import React from 'react';
+import {useAuth0} from "../auth0-wrapper";
+import Button from "./Button"
+
+const Login = () => {
+    const { isAuthenticated, loginWithRedirect, logout } = useAuth0();
+    if (!isAuthenticated) {
+        return (
+            <div>
+                <Button
+                    action={loginWithRedirect}
+                    title={'LOG IN'}
+                    type={'btn btn-primary'}>
+                </Button>
+            </div>
+        )
+    }
+    else {
+        return (
+            <div>
+                <Button
+                    action={logout}
+                    title={'LOG OUT'}
+                    type={'btn btn-danger'}>
+                </Button>
+            </div>    
+        )
+    }
+}
+
+export default Login;
