+++ smart_dog_app/src/components/PrivateRoute.jsx
