--- smart_dog_app/src/components/about.js
@@ -1,11 +1,0 @@
-import React, { Component } from 'react';
-
-class About extends Component {
-    render() {
-        return(
-            <div><h1>About Page</h1></div>
-        )
-    }
-}
-
-export default About;
