+++ smart_dog_app/src/containers/AboutPage.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const About = () => {
+    return (
+        <div><h1>About Page</h1></div>
+    )
+}
+
+export default About;
