+++ smart_dog_app/src/containers/LandingPage.jsx
