+++ smart_dog_app/src/main.js
@@ -1,16 +1,14 @@
 import React from 'react';
-import LandingPage from './landingpage';
-import About from './about';
-import Works from './works';
-import Track from './track';
-import PrivateRoute from './Private';
-//import SignUp from "./containers/SignUpForm.js"
-
+import LandingPage from './containers/LandingPage';
+import About from './containers/AboutPage';
+import Works from './containers/HowItWorksPage';
+import Track from './containers/TrackPage';
+import PrivateRoute from './components/PrivateRoute';
 import { Switch, Route } from 'react-router-dom';
 
 
 const Main = () => (
-    <Switch>
+    <Switch>W
         <Route exact path ="/" component= {LandingPage} />
         <Route path ="/about" component= {About} />
         <Route path ="/works" component= {Works} />
